Semaphores sycalls & test_sync

+++ Userland/SampleCodeModule/tests/test_processes.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <test_syscalls.h>
 #include <test_util.h>
+#include <tests.h>
 
 enum State
 {
@@ -15,8 +16,8 @@ typedef struct P_rq
 	enum State state;
 } p_rq;
 
-int64_t
-test_processes(uint64_t argc, char* argv[])
+int
+test_processes(int argc, char* argv[])
 {
 	uint8_t rq;
 	uint8_t alive = 0;
